add method ReportedVulnerability.FixIsReleased
I also moved type ReportedVulnerability and the overall report type into
their own file because more stuff will be added here. I also made the
overall report type public since it's part of the API surface of this
package anyway.
