+++ internal/trivy/trivy.go
@@ -43,20 +43,6 @@ type Config struct {
 	Token string
 }
 
-// ReportedVulnerability contains selected fields for the
-// `Results.Vulnerabilities[]` field of a Trivy vulnerability report.
-type ReportedVulnerability struct {
-	VulnerabilityID string `json:"VulnerabilityID"` // e.g. "CVE-2011-3374"
-	Severity        string `json:"Severity"`        // e.g. "HIGH", cf. clair.MapToTrivySeverity
-	FixedVersion    string `json:"FixedVersion"`    // e.g. "65.5.1"
-}
-
-type parsedTrivyReport struct {
-	Results []struct {
-		Vulnerabilities []ReportedVulnerability `json:"Vulnerabilities"`
-	} `json:"Results"`
-}
-
 // ScanManifest queries the Trivy server for a report on the given manifest.
 func (tc *Config) ScanManifest(ctx context.Context, keppelToken string, manifestRef models.ImageReference, format string) ([]byte, error) {
 	requestURL := tc.URL
@@ -92,13 +78,13 @@ func (tc *Config) ScanManifest(ctx context.Context, keppelToken string, manifest
 
 // ScanManifest is like ScanManifestAndParse, except that the result is parsed
 // instead of being returned as a bytestring.
-func (tc *Config) ScanManifestAndParse(ctx context.Context, keppelToken string, manifestRef models.ImageReference, format string) (parsedTrivyReport, error) {
+func (tc *Config) ScanManifestAndParse(ctx context.Context, keppelToken string, manifestRef models.ImageReference, format string) (VulnerabilityReport, error) {
 	report, err := tc.ScanManifest(ctx, keppelToken, manifestRef, format)
 	if err != nil {
-		return parsedTrivyReport{}, err
+		return VulnerabilityReport{}, err
 	}
 
-	var parsedReport parsedTrivyReport
+	var parsedReport VulnerabilityReport
 	err = json.Unmarshal(report, &parsedReport)
 	return parsedReport, err
 }
